fix unitest
